+++ R/spline_b.R
@@ -136,21 +136,20 @@ step_spline_b_new <-
 prep.step_spline_b <- function(x, training, info = NULL, ...) {
   col_names <- recipes_eval_select(x$terms, training, info)
   check_type(training[, col_names], types = c("double", "integer"))
+  
+  res <- list()
 
-  res <-
-    purrr::map2(
-      training[, col_names],
-      col_names,
-      ~ spline2_create(
-        .x,
-        nm = .y,
-        .fn = "bSpline",
-        df = x$deg_free,
-        complete_set = x$complete_set,
-        degree = x$degree,
-        fn_opts = x$options
-      )
+  for (col_name in col_names) {
+    res[[col_name]] <- spline2_create(
+      training[[col_name]],
+      nm = col_name,
+      .fn = "bSpline",
+      df = x$deg_free,
+      complete_set = x$complete_set,
+      degree = x$degree,
+      fn_opts = x$options
     )
+  }
   # check for errors
   bas_res <- purrr::map_lgl(res, is.null)
   res <- res[!bas_res]